Nov 07 2023
**Sum of Upper and Lower Triangle**
Given a square matrix of size N*N, print the sum of upper and lower triangular elements. Upper Triangle consists of elements on the diagonal and above it. The lower triangle consists of elements on the diagonal and below it.

Input:
N = 3
mat[][] = {{6, 5, 4},
           {1, 2, 5}
           {7, 9, 7}}
Output:
29 32
Explanation:
The given matrix is
6 5 4
1 2 5
7 9 7
The elements of upper triangle are
6 5 4
  2 5
    7
Sum of these elements is 6+5+4+2+5+7=29.
The elements of lower triangle are
6
1 2
7 9 7
Sum of these elements is 6+1+2+7+9+7= 32.
https://www.geeksforgeeks.org/problems/sum-of-upper-and-lower-triangles-1587115621/1